The arrange tools should now arrange images in the order they are in in the layer stack so images at the top will be first and images at the bottom will be last this happens only when you box select the images. You can click the same arrange button again after you just clicked it to arrange it in the opposite order.
If you want to order specific images to be first then you can manually click on each images you want to be first and the images you selected last will be at the bottom when you arrange. 

You can also export the layers to a zip file in the exact order they're in from the layer stack so the very first image starts at 1.

Updated the readme file to show you the feature changes.