feat: Fix image pasting from clipboard
This commit fixes a bug that prevented images from being pasted from the clipboard onto the canvas.

The issue was caused by a 'keydown' event listener that was unconditionally calling 'e.preventDefault()' on 'Ctrl+V', which prevented the native 'paste' event from firing.

The fix is to make the 'keydown' handler conditional, so that it only calls 'e.preventDefault()' if there are annotations on the internal clipboard. This allows the native 'paste' event to be handled by the image paste handler when the internal clipboard is empty.